requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
fastapi==0.115.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
passlib==1.7.4
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.17
python-telegram-bot==21.7
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.32.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==0.24.0
websockets==14.1
yarl==1.17.1
